Fix connection drop after 60 sec when executing commands inside containers
By default, when watching container logs using "kubectl logs -f" or
executing a command inside a container, the connection is terminated
after 60 seconds, resulting in errors like "error: unexpected EOF".
This happens because the AWS ELB in front of the k8s api server has the
"IdleTimeout" parameter set to 60. But can be increased up to 3600
seconds.

Here is the official documentation:
http://docs.aws.amazon.com/elasticloadbalancing/latest/classic/config-id
le-timeout.html