Add additional EBS volumes to worker nodes.
This adds a new feature that allows additional
EBS volumes to be deployed and mounted on all specified worker nodes.
The template in cluster.yaml follows:
```
volumeMounts:
- type: "gp2"
  iops: 0
  size: 30
  device: "/dev/xvdf"
  path: "/ebs"
```
Allowing the configuration of various parameters (which are validated).

Systemd .mount and .service resources are used to format the disks and
mount them during runtime.

NOTE: The decision to not reformat the RootVolume struct (and create a
new "VolumeMount" struct) was made to avoid issues regarding the naming
convention of "rootVolumeSize", "rootVolumeType" ... etc. already
implemented in the cluster .yaml.

Why have this feature?
The need for additional volumes arose when we realised we didn't
want a situation where the root volume would reach it's maximum capacity.