Add apiserver-count parameter in kube-apiserver config
Issues concerning conflicts between apiservers during leader allocation causes the (kubectl) kubernetes.service endpoint value to change arbitraritly. The solution implemented sets the apiserver-count parameter to the number of controllers used in the cluster.

The apiserver-count value is set to the minimum number of controller nodes set in cluster.yaml (if controller auto-scaling is used).