Follow-up for the multi API endpoints support This fixes the issue which prevented a k8s cluster from being properly configured when multiple API endpoints are defined in cluster.yaml.
It was due to empty API endpoint URLs embedded across userdata and kubeconfig.
AdminAPIEndpointURL and APIEndpointURL are introduced and injected to kubeconfig and cloud-config-worker respectively to fix the empty URLs issue.
In order to implement AdminAPIEndpointURL, the new setting `adminAPIEndpointName` in cluster.yaml is added and now required in addition to `apiEndpoints` and `worker.apiEndpointName`

Resolves #527