Ensure TLSBootstrapToken related code is not written if TLSBootstrap is disabled.
Entries that use the notation {{ if .AssetsConfig.HasAnyAuthTokens }} always return true because a TLSBootstrapToken is auto-generated by default.

If there are no AuthTokens and  TLSBootstrap is disabled then the statement {{ if .AssetsConfig.HasAnyAuthTokens }}  still returns true. Consequently, code is included in the the config-controller file which causes kube-apiserver failures (it cannot find files related to tokens).

To avoid this issue we should ensure the existence of a TLSBootstrapToken is ignored in checks if TLSBootstrap is disabled.