Allow userdata to be split across multiple parts
This change is a pure refactoring of userdata handling with zero
user-visible changes.

Userdata now is a multipart template, each part can be processed
independently. In the current form parts can be templated into
instance UserData and uploaded to S3 bucket.

2 new template functions are available to userdata templates:

 `extra`: caller can pass in place extra context, which then acessible
  to userdata as a result of `extra` function (see Etcd userdata as example)
  `self`: different part of userdata template can refer to UserData object
  using `self` function. This makes templats completely self-aware

This PR also makes userdata handling to be better integrated with
cfnstack assets management. Asset management is responsible for storing assets
on S3 and therefore acts as a source of truth for all names, S3 URLs etc. Once
part is added to Asset manager via `AddUserDataPart`, `Asset` field on part is
then set to contain infromation about location of the given part, it's final
rendered content etc.