Improved node drainer (#674)
Introduces a rewrite of the node draining logic (based on ASG lifecycle hooks) + a fix for a bug in RBAC-enabled clusters that causes pods not to be evicted (see #673).

Now, if the node drainer is enabled, the install script creates a DaemonSet that executes a control loop implemented in bash that checks the current state of each instance and, if the auto scaling group signals some instance must be terminated, the script will detect the change and will perform the node draining - with retries in case the drain command takes too long to finish.

The old node draining systemd unit was removed, as they don't really work in practice: since the node draining could take more than a few seconds to finish, it don't usually have time to finish draining the node before AWS halts the instance. (In case the operator wants to terminate some arbitrary instance, it's up to the operator to drain the node before terminating it)

Fixes #673

## Full changelog

* Add new configuration to the node drainer feature

The first configuration is to allow us to control the timeout for the
`kubectl drain` command.

The other configuration is to prevent the node from being terminated
for some time, so all drained pods could be rescheduled and started
elsewhere.

* Fix role `apiGroups`

This was causing the node draining to not work as expected when RBAC
is enabled. Without this, the Kubelet manages to cordon the node, but
not to evict pods.

* Add validation for the node drain timeout/interval

* Drain interval implementation based on ASG lifecycle hooks

This commit configures a lifecycle hook to worker auto scaling groups
and a bash script that watches the current instance status: if the
instance is terminating, we run the node drainer script and let the
lifecycle hook time out, which causes the auto scaling group to
proceed to the next instance.

I could not find a way to set up a lifecycle hook without specifying a
SNS topic since apparently the `RoleARN` and `NotificationTargetARN`
attributes are required for the `AWS::AutoScaling::LifecycleHook`
resource.

* Fix tests

* Remove the drain interval configuration

Since the idea will be to allow instances to be terminated as soon as
it is drained, there's no need for this config anymore.

* Implement the node drainer script as a DaemonSet

* Remove the old drainer system units

The idea is that voluntary node terminations should be preceded by a
`kubectl drain` executed by the cluster operator, and not leave this
to be done at the time the instance is terminated.

The main reason for this is that AWS do not wait for the drain command
to finish (especially if there are pod disruption budgets in place),
which renders this unit useless.

* Ignore node drainer configuration for node pools

Since the node drainer now runs as a DaemonSet, it should not be
possible to switch the node drainer on or off individually in each
node pool.

* Move hook target to root stack + dedicated IAM roles

* Add some comments to clarify the intent behind the changes