Send Journald logs to AWS CloudWatch
A service has been introduced which runs a dockerized image of journald-cloudwatch-logs. This service forwards journald logs to AWS CloudWatch to a LogGroup with the name of .ClusterName and is run on all nodes (Etcds, Controllers and Workers).

journald-cloudwatch-logs is a goLang project https://github.com/saymedia/journald-cloudwatch-logs.

This feature is enabled by default (with a retention period of 7 days) and configurable in cluster.yaml:

```
cloudWatchLogging:
 enabled: true
 imageWithTag: jollinshead/journald-cloudwatch-logs:0.1
 retentionInDays: 7
```