Refactor node drainer implementation
This commit separates the node drainer into two components: the ASG
watcher, and the node drainer itself.

The ASG watcher is the component (that runs as a Deployment)
responsible for using the AWS AutoScaling API to check whether there
are any instances waiting for termination; if there is, this component
annotates the corresponding node to mark it for termination.

The node drainer itself is the component (that runs as a DaemonSet)
responsible for checking each node for the termination annotation,
and draining the marked nodes.

This design allows us to significantly reduce the AutoScaling API
usage, which is a good idea, especially for large clusters.