Fix the scheduling and permissions issue of CA by fixing the conditional in stack-template.json for worker and controller stacks, while making configuration easier.
Resolves #897

The intended way of enabling cluster-autoscaler on your cluster has been to:

1. Set `addons.clusterAutoscaler.enabled` to `true`
2. Set `experimental.clusterAutoscalerSupport.enabled` to `true` to allow CA to be run on controller nodes, while providing appropriate IAM permissions to the controller nodes
3. Set `worker.nodePools[].clusterAutoscalerSupport.enabled` to `true` to allow CA to be run on worker nodes, while providing appropriate IAM permissions to the worker nodes
4. Finally, set `worker.nodePools[].autoscaling.clusterAutoscaler.enabled` to `true` for registering the node pool(s) as the scaling targets

Note that:

* 1. is required in order to deploy CA on your cluster.
* 2. and/or 3. is required to actually allow scheduling CA to any node
* 4. is required to instruct CA to start auto-scaling. Btw, missing this step had been resulted in #800 before.

`worker.nodePools[].clusterAutoscalerSupport.enabled` has been configured CA to run on worker nodes, certainly. However, it was not providing required IAM permissions to the worker nodes.
More concretely, you had to set `worker.nodePools[].addons.clusterAutoscaler.enabled` instead and that's not an expected way of configuring things. This is now fixed.

To make this process easier, `experimental.clusterAutoscalerSupport.enabled` is now set to `true` by default.
This way, 2. and 3. can be completely omitted when you don't mind on which node CA runs.

A validation to prevent a misconfiguration is also added.