Add support for CloudFormation service role
An optional setting named `cloudformation.roleARN` is added to `cluster.yaml` to instruct CloudFormation to use the specific IAM role to manage the IAM policy associated to it.

This is useful when your own IAM role is not open enough to access all the APIs and resources required by a stack operation.

Resolves #1082