Add branch to build version when not a tagged release (#1368)
When doing development/testing of kube-aws the inclusion of the branch in the kube-aws version will be a great help in quickly identifying the binary you are working with.  Now that we display the kube-aws version in the motd this also helps us confirm the geneology of an existing running cluster.

Tag releases unchange - this is only added for untagged developement builds.