Fix for rpc-statd.service needed to mount NFS file shares (#1365)
The rpc-statd.service is only startet if an EFS share is specified in the cluster.yaml. The rpc-statd.service is needed to successfully mount NFS file shares as part of POD specs. So without a EFS file share specified in the cluster.yaml you would not be able to munt any other not EFS NFS shares in your PODs.

Example that would not work without this fix

```
apiVersion: extensions/v1beta1
kind: Deployment
metadata:
  name: nfsmount
spec:
  replicas: 1
  template:
    metadata:
      labels:
        app: nfsmount
    spec:
      volumes:            
      - name: nfsserver
        nfs:
          server: my.nfsserver.host
          path: /vol1
      containers:
      - name: busybox
        image: busybox
        volumeMounts:                        
        - mountPath: /vol1
          name: nfsserver 
```

Fixes #1364