Add kube-aws:version to stacktags (#1409)
We don't currently tag our stack or instances with the version of kube-aws that created them.  This change adds a 'kube-aws:version' tag into each stack, including the root stack, and fixes a bug where tags are not updated when stacks are updated.  These stacktags are also inherited by instances, so all ec2 instances also recieve the tag.  This makes it very easy to use AWS API/CLI/GUI to check on the version of kube-aws used to deploy a cluster.

This idea was mentioned in #1405 where it was suggested as a good idea - giving us less reason for further requests to embed it the cluster.yaml.

Changelog:

* Add kube-aws/version label to stacks and instances (excludes root stack)

* Move kube-aws:version into StackTags and fix tags root stack tags not being updated on cluster updates

* Make the stacktags map if it is empty - as it is in testing