faet: `iam.role.useStrict` to share existing iam role for controller nodes (#1433)
## What

This feature allows the user to specify if they want to use the `IAM.role.name` as strict (without appending the region and cluster-name to it) by providing a new flag under `IAM.role` in the `cluster.yaml` configuration file. E.g:
```yaml
controller:
  iam:
    role:
       name: myrole1
       useStrict: true
```

When the cluster is rolled-out with the above example, `kube-aws` will not create a new role (e.g `<clusterName> - <region> - myrole1 ` ) but will instead use `myrole1` that already exists on AWS as an `IAM Role`.

The `IAM role` needs to already be created and have a policy attached. What this feature does is simply assign the existing role (specified in the cluster.yaml file) to the `IAMControllerInstanceProfile` that is created upon cluster roll-outs, to replace the `IAMControllerRole` that is created along with the instance profile.

## Why

In order to create your own [Kiam](https://github.com/uswitch/kiam) roles you need to add a trust. The trust is part of the `Controller` and only gets created when the cluster is rolled-out. At Hotels.com, we would like to manage our `Kiam` specific roles and policies before we roll-out our cluster, and possibly re-use the same role for multiple cluster in the same AWS account. 

Currently, we cannot manage roles prior to rolling out our cluster because they cannot point to the Kiam specific trust, in order to point to it we would have to wait for it to be available and for the cluster to roll out. This keeps the role name predictable.

## How

In order to have an existing role attached to the ControllerInstanceProfile, we need to have that role's name. To keep the default behaviour, this feature uses the `IAM.Role.Name` field and introduces an extra field under `IAM.Role` called `UseStrict` (to use the strict form of the role name with nothing appended or prepended).

* By default, `UseStrict` is set to false. This keeps the default behaviour and the change will not affect existing clusters.

* If enabled, instead of creating a new role (`IAMControllerRole`) and then adding that to the `IAMInstanceProfileController`,  we simply pass the name of the existing role (taken from `IAM.Role.Name`) to the `IAMInstanceProfileController`. 

* If it is disabled, then we create a new controller role.

___Note:___ the same can be achieved by manually removing the role from the `IAMInstanceProfileController`, adding our existing role to the instance profile and then deleting the kiam server pods using the kube api, this however takes long and does not fix the initial issue of managing roles  before rolling out a cluster and minimising loss in the pods that use Kiam policies.

Changelog:

* Added IAM.ManagedRole for InstanceProfile

* Added ManagedRole to use as a pre-existing role when deploying

* Added assertion for managedRole in controller.iam

* Changed stack template to accommodate new field and logic

Documented the use of useStrict

Changed managed to be useStrict and moved it under role

Rewrote the default test and added another one in