Cannot depend on resources in another stack (#1670)
The dependency on the other stack gives errors. Starting the etcd node
before the natgateway is deployed is not a big issue as it keeps
restarting its proces to fetch its data. Wait signal should be long
enough for this to be resolved.

Fixes https://github.com/kubernetes-incubator/kube-aws/issues/1332