Allow dnsmasq to be backed by a local copy of CoreDNS
This commit introduces a new configuration option for cluster.yaml,
`kubeDns.coreDNSLocal`. If this option and `kubeDns.nodeLocalResolver`
are set to true, the dnsmasq-node DaemonSet will be configured to use
a local copy of CoreDNS for its resolution while setting the global
CoreDNS service as a fallback. This is handy in situations where the
number of DNS requests within a cluster grows large and causes
resolution issues as dnsmasq reaches out to the global CoreDNS
service.

See [0] for an investigation into this situation which was instrumental
in understanding issues we were facing. Many thanks to dominicgunn for providing
the manifests which I codified into this commit.

0: https://github.com/zalando-incubator/kubernetes-on-aws/blob/dev/docs/postmortems/jan-2019-dns-outage.md