Increase DNS autosacaler memory and add limit
## Problem
When in large clusters memory allocation can be a challenge if `memory limit != memory requests`. Because of that we've some internal requirements for the platform and we're committing to be more aggressive and taking this very seriously in our platform.
We've some tooling around this in prometheus and as part of our memory allocation cleanup process we've found kube-dns-autoscaler consuming 250% times the requested memory. As we've this alert in our side, this component will trigger an alert: 
```
(
  avg(
    container_memory_usage_bytes{
      container!="POD",
      image!="",
    }
  ) by (container, namespace) 
  * 100
) / max(
  kube_pod_container_resource_requests_memory_bytes{
    job="kube-state-metrics",
    }
  ) by (container, namespace) > 90
```

## Solution

I've increased the memory request up to 50Mi and added a proper memory limit.

### Note

This cloud be improved by just adding parameters in the cluster.yaml but I didn't take the time as I think not so much people will face this issue