Always create dnsmasq-node-coredns-local.yaml
If this file does not exist (as would be the case if the CoreDNS local
feature has not been enabled), controller nodes will fail to come up
with the error:
> error: the path "/srv/kubernetes/manifests/dnsmasq-node-coredns-local.yaml" does not exist
This is caused when `kubectl delete` is called against the file because
of the line `remove "${mfdir}/dnsmasq-node-coredns-local.yaml`.

This manifest must always be generated because the CoreDNS-local
feature cannot be enabled and then later disabled without otherwise
requiring manual operator intervention.